modified : add 42 haeder to functions

+++ src/bzero/ft_bzero.c
@@ -1,10 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_bzero.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mnchabel <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2019/05/27 11:28:48 by mnchabel          #+#    #+#             */
+/*   Updated: 2019/05/27 11:28:50 by mnchabel         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include <string.h>
 #include <stdio.h>
 
-void	ft_bzero(void *s, int n)
+void	ft_bzero(void *s, size_t n)
 {
 	unsigned char	*str;
-	int				i;
+	size_t			i;
 
 	str = (unsigned char *)s;
 	i = 0;
@@ -19,9 +31,9 @@ int	main(void)
 {
 	char str[20];
 
-        bzero(str, 1);
-        printf("bzero : %s\n", str);
-        ft_bzero(str, 1);
-        printf("ft_bzero : %s\n", str);
+	bzero(str, 1);
+	printf("bzero : %s\n", str);
+	ft_bzero(str, 1);
+	printf("ft_bzero : %s\n", str);
 	return (0);
 }
