+++ src/memcmp/ft_memcmp.c
@@ -1,11 +1,23 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memcmp.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mnchabel <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2019/05/27 10:18:02 by mnchabel          #+#    #+#             */
+/*   Updated: 2019/05/27 10:18:13 by mnchabel         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include <string.h>
 #include <stdio.h>
 
-int	ft_memcmp(const void *s1, const void *s2, int n)
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	unsigned char	*ft_s1;
 	unsigned char	*ft_s2;
-	int				i;
+	size_t			i;
 
 	if (s1 == s2 || n == 0 || (!s1 && !s2 && !n))
 		return (0);
