+++ src/memcpy/ft_memcpy.c
@@ -1,11 +1,23 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memcpy.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mnchabel <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2019/05/27 11:33:34 by mnchabel          #+#    #+#             */
+/*   Updated: 2019/05/27 11:33:36 by mnchabel         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include <string.h>
 #include <stdio.h>
 
-void	*ft_memcpy(void *dest, const void *src, int n)
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	unsigned char	*str_dest;
 	unsigned char	*str_src;
-	int				i;
+	size_t			i;
 
 	str_dest = (unsigned char *)dest;
 	str_src = (unsigned char *)src;
