+++ src/memmove/ft_memmove.c
@@ -1,11 +1,23 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memmove.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mnchabel <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2019/05/27 11:33:49 by mnchabel          #+#    #+#             */
+/*   Updated: 2019/05/27 11:33:51 by mnchabel         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include <string.h>
 #include <stdio.h>
 
-void	*ft_memmove(void *dest, void *src, int len)
+void	*ft_memmove(void *dest, void *src, size_t len)
 {
 	unsigned char	*ft_dest;
 	unsigned char 	*ft_src;
-	int				i;
+	size_t			i;
 
 	ft_dest = (unsigned char *)dest;
 	ft_src = (unsigned char *)src;
