+++ src/strlen/ft_strlen.c
@@ -1,9 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strlen.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mnchabel <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2019/05/27 10:21:02 by mnchabel          #+#    #+#             */
+/*   Updated: 2019/05/27 10:23:16 by mnchabel         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include <string.h>
 #include <stdio.h>
 
-unsigned int	ft_strlen(char *s)
+size_t	ft_strlen(char *s)
 {
-	unsigned int i;
+	size_t i;
 
 	i = 0;
 	while (s[i])
